Refactor vehicles api
