+++ backend/vehicles/src/app.module.ts
@@ -8,10 +8,10 @@ import * as path from 'path';
 import { AppController } from './app.controller';
 import { AutomapperModule } from '@automapper/nestjs';
 import { classes } from '@automapper/classes';
-import { PowerUnitTypesModule } from './modules/power-unit-types/power-unit-types.module';
-import { PowerUnitsModule } from './modules/power-units/power-units.module';
-import { TrailerTypesModule } from './modules/trailer-types/trailer-types.module';
-import { TrailersModule } from './modules/trailers/trailers.module';
+import { PowerUnitTypesModule } from './modules/vehicles/power-unit-types/power-unit-types.module';
+import { PowerUnitsModule } from './modules/vehicles/power-units/power-units.module';
+import { TrailerTypesModule } from './modules/vehicles/trailer-types/trailer-types.module';
+import { TrailersModule } from './modules/vehicles/trailers/trailers.module';
 
 const envPath = path.resolve(process.cwd() + '/../../');
