+++ backend/vehicles/src/modules/vehicles/power-unit-types/dto/request/create-power-unit-type.dto.ts
