+++ backend/vehicles/src/modules/vehicles/power-unit-types/dto/request/update-power-unit-type.dto.ts
