+++ backend/vehicles/src/modules/vehicles/power-unit-types/dto/response/read-power-unit-type.dto.ts
