+++ backend/vehicles/src/modules/vehicles/power-unit-types/power-unit-types.controller.ts
@@ -19,8 +19,8 @@ import {
   ApiTags,
 } from '@nestjs/swagger';
 import { ReadPowerUnitTypeDto } from './dto/response/read-power-unit-type.dto';
-import { ExceptionDto } from '../../common/dto/exception.dto';
-import { DataNotFoundException } from '../../common/exception/data-not-found.exception';
+import { ExceptionDto } from '../../../common/dto/exception.dto';
+import { DataNotFoundException } from '../../../common/exception/data-not-found.exception';
 
 @ApiTags('Power Unit Types')
 @ApiNotFoundResponse({
