+++ backend/vehicles/src/modules/vehicles/power-unit-types/power-unit-types.service.ts
