+++ backend/vehicles/src/modules/vehicles/power-unit-types/profiles/power-unit-type.profile.ts
