+++ backend/vehicles/src/modules/vehicles/power-units/profiles/power-unit.profile.ts
