+++ backend/vehicles/src/modules/vehicles/trailer-types/dto/request/create-trailer-type.dto.ts
