+++ backend/vehicles/src/modules/vehicles/trailer-types/dto/request/update-trailer-type.dto.ts
