+++ backend/vehicles/src/modules/vehicles/trailer-types/dto/response/read-trailer-type.dto.ts
