+++ backend/vehicles/src/modules/vehicles/trailer-types/entities/trailer-type.entity.ts
@@ -1,32 +1,32 @@
 import { AutoMap } from '@automapper/classes';
 import { ApiProperty } from '@nestjs/swagger';
 import { Entity, Column, OneToMany, PrimaryColumn } from 'typeorm';
-import { Base } from '../../../common/entities/base.entity';
-import { PowerUnit } from '../../power-units/entities/power-unit.entity';
+import { Base } from '../../../../common/entities/base.entity';
 
-@Entity({ name: 'ORBC_VT_POWER_UNIT_TYPE' })
-export class PowerUnitType extends Base {
+import { Trailer } from '../../trailers/entities/trailer.entity';
+
+@Entity({ name: 'ORBC_VT_TRAILER_TYPE' })
+export class TrailerType extends Base {
   @AutoMap()
   @ApiProperty({
-    example: 'CONCRET',
-    description: 'Unique identifier of the power unit type.',
+    example: 'BOOSTER',
+    description: 'Unique identifier of the trailer type.',
   })
   @PrimaryColumn({ length: 7, name: 'TYPE_CODE', nullable: false })
   typeCode: string;
 
   @AutoMap()
   @ApiProperty({
-    example: 'Concrete Pumper Trucks',
-    description: 'Short description of the power unit type.',
+    example: 'Boosters',
+    description: 'Short description of the trailer type.',
   })
   @Column({ length: 150, name: 'TYPE', nullable: false })
   type: string;
 
   @AutoMap()
   @ApiProperty({
-    example:
-      'Concrete Pumper Trucks are used to pump concrete from a cement mixer truck to where the concrete is actually needed. They travel on the highway at their equipped weight with no load.',
-    description: 'Long description of the power unit type.',
+    example: 'A Booster is similar to a jeep, but it is used behind a load.',
+    description: 'Long description of the trailer type.',
   })
   @Column({ length: 500, name: 'DESCRIPTION', nullable: true })
   description: string;
@@ -47,7 +47,7 @@ export class PowerUnitType extends Base {
   @Column({ type: 'bit', name: 'IS_ACTIVE', nullable: false, default: 1 })
   isActive: string;
 
-  @AutoMap(() => [PowerUnit])
-  @OneToMany(() => PowerUnit, (PowerUnit) => PowerUnit.powerUnitType)
-  powerUnits: PowerUnit[];
+  @AutoMap(() => [Trailer])
+  @OneToMany(() => Trailer, (Trailer) => Trailer.trailerType)
+  trailers: Trailer[];
 }
