+++ backend/vehicles/src/modules/vehicles/trailer-types/profiles/trailer-type.profile.ts
