+++ backend/vehicles/src/modules/vehicles/trailer-types/trailer-types.service.ts
