+++ backend/vehicles/src/modules/vehicles/trailers/dto/request/create-trailer.dto.ts
