+++ backend/vehicles/src/modules/vehicles/trailers/dto/request/update-trailer.dto.ts
