+++ backend/vehicles/src/modules/vehicles/trailers/dto/response/read-trailer.dto.ts
