+++ backend/vehicles/test/util/mocks/data/power-unit.mock.ts
@@ -1,7 +1,7 @@
-import { CreatePowerUnitDto } from '../../../../src/modules/power-units/dto/request/create-power-unit.dto';
-import { UpdatePowerUnitDto } from '../../../../src/modules/power-units/dto/request/update-power-unit.dto';
-import { ReadPowerUnitDto } from '../../../../src/modules/power-units/dto/response/read-power-unit.dto';
-import { PowerUnit } from '../../../../src/modules/power-units/entities/power-unit.entity';
+import { CreatePowerUnitDto } from '../../../../src/modules/vehicles/power-units/dto/request/create-power-unit.dto';
+import { UpdatePowerUnitDto } from '../../../../src/modules/vehicles/power-units/dto/request/update-power-unit.dto';
+import { ReadPowerUnitDto } from '../../../../src/modules/vehicles/power-units/dto/response/read-power-unit.dto';
+import { PowerUnit } from '../../../../src/modules/vehicles/power-units/entities/power-unit.entity';
 import { baseDtoMock, baseEntityMock } from './base.mock';
 import { powerUnitTypeEntityMock } from './power-unit-type.mock';
 import { provinceEntityMock } from './province.mock';
