+++ backend/vehicles/test/util/mocks/data/trailer-type.mock.ts
@@ -1,7 +1,7 @@
-import { CreateTrailerTypeDto } from '../../../../src/modules/trailer-types/dto/request/create-trailer-type.dto';
-import { UpdateTrailerTypeDto } from '../../../../src/modules/trailer-types/dto/request/update-trailer-type.dto';
-import { ReadTrailerTypeDto } from '../../../../src/modules/trailer-types/dto/response/read-trailer-type.dto';
-import { TrailerType } from '../../../../src/modules/trailer-types/entities/trailer-type.entity';
+import { CreateTrailerTypeDto } from '../../../../src/modules/vehicles/trailer-types/dto/request/create-trailer-type.dto';
+import { UpdateTrailerTypeDto } from '../../../../src/modules/vehicles/trailer-types/dto/request/update-trailer-type.dto';
+import { ReadTrailerTypeDto } from '../../../../src/modules/vehicles/trailer-types/dto/response/read-trailer-type.dto';
+import { TrailerType } from '../../../../src/modules/vehicles/trailer-types/entities/trailer-type.entity';
 import { baseEntityMock } from './base.mock';
 
 const TYPE_CODE = 'BOOSTER';
